Delete cloudbase.yml
Getting rid of this since it (a) doesn't work yet and (b) isn't really what we want for our workflow. What makes more sense is to have a dev server and a live server and control deployment to live manually (since it resets the state and will soon be in use for testing pilot instrument configurations).